chore: remove helium run loop infrastructure
