+++ stacks-node/src/run_loop/nakamoto.rs
@@ -41,7 +41,7 @@ use crate::globals::Globals as GenericGlobals;
 use crate::monitoring::{start_serving_monitoring_metrics, MonitoringError};
 use crate::nakamoto_node::{self, StacksNode, BLOCK_PROCESSOR_STACK_SIZE, RELAYER_MAX_BUFFER};
 use crate::neon_node::LeaderKeyRegistrationState;
-use crate::node::{
+use crate::genesis::{
     get_account_balances, get_account_lockups, get_names, get_namespaces,
     use_test_genesis_chainstate,
 };
